# docs/boss_swagger.yml
swagger: "2.0"
info:
  description: "This is a sample BOSS server."
  version: "1.0.0"
  title: "Swagger BOSS"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "h.cankaynak@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "petstore.swagger.io"
basePath: "/v2"
tags:
  - name: "user"
    description: "Everything about User"
    externalDocs:
      description: "Find out more"
      url: "http://swagger.io"
  - name: "store"
    description: "Access to Petstore orders"
  - name: "user"
    description: "Operations about user"
    externalDocs:
      description: "Find out more about our store"
      url: "http://swagger.io"
schemes:
  - "https"
  - "http"
paths:
  /users/signUp:
    post:
      tags:
        - "user"
      summary: "Create user"
      description: "This is basic sign Up"
      operationId: "createUser"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Created user object"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        default:
          description: "successful operation"
  /users/login:
    post:
      tags:
        - "user"
      summary: "Login"
      description: "This is basic login"
      operationId: "login"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Created user object"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        default:
          description: "successful operation"

  /category:
    post:
      tags:
        - "category"
      summary: "create a category"
      description: "This is basic create function"
      operationId: "createCategory"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "create category"
          required: true
          schema:
            $ref: "#/definitions/Category"
      responses:
        default:
          description: "successful operation"
  /products:
    post:
      tags:
        - "user"
      summary: "Product Add"
      description: "This is basic product adder"
      operationId: "productAdd"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Created product object"
          required: true
          schema:
            $ref: "#/definitions/Product"
      responses:
        default:
          description: "successful operation"

  /basket:
    post:
      tags:
        - "basket"
      summary: "Product Add To Basket"
      description: "This is basic product adder To basket"
      operationId: "productAddToBasket"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Add Product To Basket"
          required: true
          schema:
            $ref: "#/definitions/AddToBasket"
      responses:
        default:
          description: "successful operation"
    get:
      tags:
        - "basket"
      summary: "Get Basket"
      description: "This is basic product get basket"
      operationId: "getBasket"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Get Basket"
          required: true
          schema:
            $ref: "#/definitions/Product"
      responses:
        default:
          description: "successful operation"

  /category/list:
    get:
      tags:
        - "category"
      summary: "list all item in category if active and not deleted"
      description: "This is basic listing function"
      operationId: "listCategory"
      produces:
        - "application/json"
      parameters:
        - name: "page"
          in: query
          type: integer
          description: "page"
          required: true
        - name: "pageSize"
          in: query
          description: "page size"
          required: true
          type: integer
      responses:
        default:
          description: "successful operation"

securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "uint"
      email:
        type: "string"
      password:
        type: "string"
      name:
        type: "string"
      role:
        type: "string"
    xml:
      name: "User"
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
      type:
        type: "string"
      message:
        type: "string"
      details:
        description: a (key, value) map.
        type: object

  Category:
    type: "object"
    properties:
      name:
        type: "string"
      isDeleted:
        type: "boolean"
      isActive:
        type: "boolean"

  AddToBasket:
    type: "object"
    properties:
      id:
        type: "integer"

  Product:
    type: "object"
    properties:
      name:
        type: "string"
      description:
        type: "string"

externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"