# internal/api/add_to_basket.go
// Code generated by go-swagger; DO NOT EDIT.

package api

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// AddToBasket add to basket
//
// swagger:model AddToBasket
type AddToBasket struct {

	// id
	ID int64 `json:"id,omitempty"`
}

// Validate validates this add to basket
func (m *AddToBasket) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this add to basket based on context it is used
func (m *AddToBasket) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *AddToBasket) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *AddToBasket) UnmarshalBinary(b []byte) error {
	var res AddToBasket
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}